backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
motor==3.3.1
pymongo==4.5.0
pydantic==2.12.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-dotenv==1.2.1
google-cloud-secret-manager==2.18.0
httpx==0.28.1
python-multipart==0.0.20
email-validator==2.3.0
PyJWT==2.10.1
aiohttp==3.9.1
